% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_CRT.R
\name{robust_CRT}
\alias{robust_CRT}
\title{Cluster Randomized Trials with Robust Estimators}
\usage{
robust_CRT(
  formula,
  data,
  clus_id,
  trt,
  prob,
  method,
  family = gaussian(link = "identity"),
  corstr,
  scale,
  jack = 1,
  alpha = 0.05
)
}
\arguments{
\item{formula}{A formula for the outcome mean model, including covariates.}

\item{data}{A data frame where categorical variables should already be converted to dummy variables.}

\item{clus_id}{A string representing the column name of the cluster ID in the data frame.}

\item{trt}{A string representing the column name of the treatment assignment per cluster.}

\item{prob}{A vector of probabilities of treatment assignment per cluster, conditional on covariates.}

\item{method}{A string specifying the outcome mean model. Possible values are:
- 'LM': linear model on cluster-level means (continuous outcome).
- 'LMM': linear mixed model on individual-level data (continuous outcome).
- 'GEE': generalized estimating equations.
- 'GLMM': generalized linear mixed model.}

\item{family}{The link function for the outcome. Can be one of the following:
- `gaussian(link = "identity")`: for continuous outcomes. Default is gaussian("identity")
- `binomial(link = "logit")`: for binary outcomes.
- `poisson(link = "log")`: for count outcomes.}

\item{corstr}{A string specifying the correlation structure for GEE models (e.g., "exchangeable", "independence").}

\item{scale}{A string specifying the risk measure of interest. Can be 'RD' (risk difference), 'RR' (relative risk), or 'OR' (odds ratio).}

\item{jack}{A numeric value (1, 2, or 3) specifying the type of jackknife standard error estimate. Type 1 is the standard jackknife, and type 3 is recommended for small numbers of clusters. Default is 1.}

\item{alpha}{A numeric value for the type-I error rate. Default is 0.05.}
}
\value{
A list with the following components:
  - `estimate`: A summary table of estimates.
  - `m`: Number of clusters.
  - `N`: Total number of observations per cluster.
  - `family`: The family used for the model.
  - `model`: The method used for the outcome mean model.
}
\description{
This function performs cluster randomized trials (CRT) analysis using robust estimators to estimate the cluster-average and individual-average treatment effect. It handles different outcome mean models (LM, LMM, GEE, GLMM) and supports both continuous and binary outcomes with options for different correlation structures and scales (risk different, risk ratio and odds ratio).
}
\examples{
\dontrun{
ppact_prob <- ppact \%>\%
  group_by(CLUST) \%>\%
 mutate(first_trt = first(INTERVENTION)) \%>\%
 ungroup() \%>\%
 mutate(prob_A_1 = mean(first_trt == 1, na.rm = TRUE),  # Proportion trt=1
        prob_A_0 = mean(first_trt == 0, na.rm = TRUE)) \%>\%
 mutate(assigned_value = ifelse(INTERVENTION == 1, prob_A_1, prob_A_0))

prob <- ppact_prob$assigned_value

example <- robust_CRT(
  formula = PEGS ~ AGE + FEMALE + comorbid + Dep_OR_Anx + pain_count+PEGS_bl +
   BL_benzo_flag + BL_avg_daily + satisfied_primary + cluster(n),
 data = ppact,
 clus_id = "CLUST",
 trt = "INTERVENTION",
 prob = prob,
 method = "GEE",
 corstr = "independence",
 scale = "RR"
)
}
}
