% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRT_point.R
\name{robust_CRT_point}
\alias{robust_CRT_point}
\title{Robust CRT Point Estimate}
\usage{
robust_CRT_point(
  formula,
  data,
  clus_id,
  trt,
  prob,
  family = gaussian(link = "identity"),
  corstr,
  method = "LM",
  scale
)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables for the CRT analysis.}

\item{clus_id}{identify the column containing cluster level information(factor)}

\item{trt}{The variable names for treatment}

\item{prob, }{vector of probability for each subject based on cluster treatment assignment}

\item{family}{identify the outcome family for fitting the gee, glmm, etc}

\item{corstr}{correlation structure}

\item{method}{model for the outcome regression}

\item{scale}{the estimand, RR, RD, or OR}
}
\value{
A numeric value representing the point estimate for the robust CRT.
}
\description{
This function calculates a point estimate for a robust clustered randomized trial (CRT).
}
